refactor(BoardBuilder): Remove needless class RandomBoardBuilder
